fix version script
